lncfg: move ParseHexColor to lncfg
